Moving stamps from cbc to base package
