+++ base/header_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/invopop/gobl/base"
-	"github.com/invopop/gobl/cbc"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -15,18 +14,18 @@ func TestNewHeader(t *testing.T) {
 	assert.NotPanics(t, func() {
 		h.Meta["foo"] = "bar"
 		h.Tags = append(h.Tags, "foo")
-		h.Stamps = append(h.Stamps, &cbc.Stamp{})
+		h.Stamps = append(h.Stamps, &base.Stamp{})
 	}, "header and meta hash should have been initialized")
 }
 
 func TestHeaderAddStamp(t *testing.T) {
 	h := base.NewHeader()
-	h.AddStamp(&cbc.Stamp{Provider: "foo", Value: "bar"})
+	h.AddStamp(&base.Stamp{Provider: "foo", Value: "bar"})
 	assert.Len(t, h.Stamps, 1)
-	h.AddStamp(&cbc.Stamp{Provider: "foo", Value: "baz"})
+	h.AddStamp(&base.Stamp{Provider: "foo", Value: "baz"})
 	assert.Len(t, h.Stamps, 1)
 	assert.Equal(t, "baz", h.Stamps[0].Value)
-	h.AddStamp(&cbc.Stamp{Provider: "bar", Value: "bax"})
+	h.AddStamp(&base.Stamp{Provider: "bar", Value: "bax"})
 	assert.Len(t, h.Stamps, 2)
 	assert.Equal(t, "bax", h.Stamps[1].Value)
 }
