+++ base/stamps_test.go
@@ -1,8 +1,9 @@
-package cbc_test
+package base_test
 
 import (
 	"testing"
 
+	"github.com/invopop/gobl/base"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/validation"
 	"github.com/stretchr/testify/assert"
@@ -10,9 +11,9 @@ import (
 
 func TestDuplicateStamps(t *testing.T) {
 	st := struct {
-		Stamps []*cbc.Stamp
+		Stamps []*base.Stamp
 	}{
-		Stamps: []*cbc.Stamp{
+		Stamps: []*base.Stamp{
 			{
 				Provider: cbc.Key("provider"),
 				Value:    "value",
@@ -25,16 +26,16 @@ func TestDuplicateStamps(t *testing.T) {
 	}
 
 	err := validation.ValidateStruct(&st,
-		validation.Field(&st.Stamps, cbc.DetectDuplicateStamps),
+		validation.Field(&st.Stamps, base.DetectDuplicateStamps),
 	)
 	assert.NoError(t, err)
 
-	st.Stamps = append(st.Stamps, &cbc.Stamp{
+	st.Stamps = append(st.Stamps, &base.Stamp{
 		Provider: cbc.Key("provider"),
 		Value:    "value3",
 	})
 	err = validation.ValidateStruct(&st,
-		validation.Field(&st.Stamps, cbc.DetectDuplicateStamps),
+		validation.Field(&st.Stamps, base.DetectDuplicateStamps),
 	)
 	assert.Error(t, err)
 	assert.Contains(t, err.Error(), "duplicate stamp 'provider'")
@@ -42,16 +43,16 @@ func TestDuplicateStamps(t *testing.T) {
 
 func TestAddStamp(t *testing.T) {
 	st := struct {
-		Stamps []*cbc.Stamp
+		Stamps []*base.Stamp
 	}{
-		Stamps: []*cbc.Stamp{
+		Stamps: []*base.Stamp{
 			{
 				Provider: cbc.Key("provider"),
 				Value:    "value",
 			},
 		},
 	}
-	st.Stamps = cbc.AddStamp(st.Stamps, &cbc.Stamp{
+	st.Stamps = base.AddStamp(st.Stamps, &base.Stamp{
 		Provider: cbc.Key("provider"),
 		Value:    "new value",
 	})
